feat: add user-todo relationship migrations

+++ router/todos.js
@@ -6,12 +6,13 @@ const todo = db.Todo;
 
 router.get('/', (req, res, next) => {
   let { page, limit } = req.query;
+  const userId = req.user.id;
   page = parseInt(page) || 1;
   limit = parseInt(limit) || 10;
-  console.log(limit);
   return todo
     .findAndCountAll({
       attributes: ['id', 'name', 'isComplete'],
+      where: { userId },
       raw: true,
       offset: (page - 1) * limit,
       limit,
@@ -37,18 +38,28 @@ router.get('/new', (req, res, next) => {
 
 router.get('/:id', (req, res, next) => {
   const id = req.params.id;
+  const userId = req.user.id;
   return todo
     .findByPk(id, { raw: true })
-    .then((todo) => {
-      res.render('detail', { todo });
+    .then((td) => {
+      if (!td) {
+        req.flash('error', 'Todo not found!');
+        return res.redirect('/todos');
+      }
+      if (td.userId !== userId) {
+        req.flash('error', 'Permission denied!');
+        return res.redirect('/todos');
+      }
+      res.render('detail', { todo: td });
     })
     .catch((err) => console.log(err));
 });
 
 router.post('/', (req, res, next) => {
   const { isComplete, name } = req.body;
+  const userId = req.user.id;
   return todo
-    .create({ name, isComplete: isComplete ? 1 : 0 })
+    .create({ name, userId, isComplete: isComplete ? 1 : 0 })
     .then(() => {
       req.flash('success', 'Add new successfully!');
       res.redirect('/todos');
@@ -61,39 +72,72 @@ router.post('/', (req, res, next) => {
 
 router.get('/:id/edit', (req, res, next) => {
   const id = req.params.id;
+  const userId = req.user.id;
   return todo
-    .findByPk(id, { attributes: ['id', 'name', 'isComplete'], raw: true })
-    .then((todo) => res.render('edit', { todo }))
+    .findByPk(id, { raw: true })
+    .then((td) => {
+      if (!td) {
+        req.flash('error', 'Todo not found!');
+        return res.redirect('/todos');
+      }
+      if (td.userId !== userId) {
+        req.flash('error', 'Permission denied!');
+        return res.redirect('/todos');
+      }
+      res.render('edit', { todo: td });
+    })
     .catch((err) => console.log(err));
 });
 
 router.put('/:id', (req, res, next) => {
   const id = req.params.id;
+  const userId = req.user.id;
   const { name, isComplete } = req.body;
-  return todo
-    .update({ name, isComplete: isComplete ? 1 : 0 }, { where: { id } })
-    .then((result) => {
-      req.flash('success', 'Update successfully!');
-      res.redirect(`/todos/${id}`);
-    })
-    .catch((err) => {
-      err.errorMessage = 'something went wrong!';
-      next(err);
-    });
+  return todo.findByPk(id, { raw: true }).then((td) => {
+    if (!td) {
+      req.flash('error', 'Todo not found!');
+      return res.redirect('/todos');
+    }
+    if (td.userId !== userId) {
+      req.flash('error', 'Permission denied!');
+      return res.redirect('/todos');
+    }
+    todo
+      .update({ name, isComplete: isComplete ? 1 : 0 }, { where: { id } })
+      .then((result) => {
+        req.flash('success', 'Update successfully!');
+        res.redirect(`/todos/${id}`);
+      })
+      .catch((err) => {
+        err.errorMessage = 'something went wrong!';
+        next(err);
+      });
+  });
 });
 
 router.delete('/:id', (req, res, next) => {
   const id = req.params.id;
-  return todo
-    .destroy({ where: { id } })
-    .then(() => {
-      req.flash('success', 'Delete successfully!');
-      res.redirect('/todos');
-    })
-    .catch((err) => {
-      err.errorMessage = 'something went wrong!';
-      next(err);
-    });
+  const userId = req.user.id;
+  return todo.findByPk(id, { raw: true }).then((td) => {
+    if (!td) {
+      req.flash('error', 'Todo not found!');
+      return res.redirect('/todos');
+    }
+    if (td.userId !== userId) {
+      req.flash('error', 'Permission denied!');
+      return res.redirect('/todos');
+    }
+    todo
+      .destroy({ where: { id } })
+      .then(() => {
+        req.flash('success', 'Delete successfully!');
+        res.redirect('/todos');
+      })
+      .catch((err) => {
+        err.errorMessage = 'something went wrong!';
+        next(err);
+      });
+  });
 });
 
 module.exports = router;
